Fix "InvalidArgumentError: body must be a string, a Buffer or a Readable stream"
Not sure if this is a good fix so I haven't included tests yet. If it is I'll add tests to this PR :+1: .

Reproduce issue:
```js
import fetch from 'undici-fetch';

const r = await fetch('https://httpbin.org/post', {
    method: 'POST',
    body: 'test=123'
});
```

the module is passing a `ControlledAsyncIterable` to undici which causes the error.